Initial check in of my script, this updates the server (intended to be hosted on Ubuntu 16.04.4) installs steamcmd, creates the "/ark" folder and sets it to be owned by your username.
Next it sets a bunch of required items, by building files or appending
to copies of the system files (for pam & file limits) and copies these
updted versions back to the various locations.

Steam command is then run and quit, to just update it.

And then everything in the /ark folder it set to be owned by your
username, not root.

We now add the "ark" user and then link steamcmd to our home directory
for ease, as we then download the ark server application from steam (id
376030).  Once this is complete we open the Default game settings in
nano, so you can set anything and everything up.

(NOTE: this has no password for the server at this point)

We have to then revert ownership and group control of the "/ark" folder
to the ark user.

The Final step is the systemd service, we build a script for this now in
our home directory and copy it into place, update systemctl, enable the
service and then try to start ark.

Hopefully everything should start and your server should have no
password, but be called "XelsServer" - as I am xelous.

You can add a password on the service lines if you wish, and change the
port, but we'll default everything here in this script.